fix build, add depool validator script

+++ build.sh
@@ -9,7 +9,7 @@ rm -rf "${TMP_DIR}"
 mkdir -p "${TMP_DIR}"
 
 NODE_BUILD_DIR="${SRC_TOP_DIR}/build"
-TOOLS_BUILD_DIR="${SRC_TOP_DIR}/build/ton-node/ton-labs-node-tools"
+TOOLS_BUILD_DIR="${SRC_TOP_DIR}/build/ton-labs-node-tools"
 TONOS_CLI_BUILD_DIR="${SRC_TOP_DIR}/build/tonos-cli"
 
 BIN_DIR="${SRC_TOP_DIR}/bin"
@@ -48,23 +48,29 @@ cd "${NODE_BUILD_DIR}/ton-node" && git checkout "${TON_NODE_GITHUB_COMMIT_ID}"
 cargo update
 cargo build --release
 
-mv "${BIN_DIR}/ton_node" "${TMP_DIR}/" &> /dev/null || true
-cp "${NODE_BUILD_DIR}/ton-node/target/release/ton_node" "${BIN_DIR}/"
+if [ -f "${NODE_BUILD_DIR}/ton-node/target/release/ton_node" ]; then
+    mv "${BIN_DIR}/ton_node" "${TMP_DIR}/" &> /dev/null || true
+    cp "${NODE_BUILD_DIR}/ton-node/target/release/ton_node" "${BIN_DIR}/"
+fi
 
 echo "INFO: build a node... DONE"
 
 echo "INFO: build utils (ton-labs-node-tools)..."
 
-cd "${NODE_BUILD_DIR}/ton-node" && git clone --recursive "${TON_NODE_TOOLS_GITHUB_REPO}"
-cd "${NODE_BUILD_DIR}/ton-node/ton-labs-node-tools" && git checkout "${TON_NODE_TOOLS_GITHUB_COMMIT_ID}"
+cd "${NODE_BUILD_DIR}/" && git clone --recursive "${TON_NODE_TOOLS_GITHUB_REPO}"
+cd "${NODE_BUILD_DIR}/ton-labs-node-tools" && git checkout "${TON_NODE_TOOLS_GITHUB_COMMIT_ID}"
 
 cargo update
 cargo build --release
 
-mv "${TOOLS_DIR}/console" "${TMP_DIR}/" &> /dev/null || true
-cp "${NODE_BUILD_DIR}/ton-node/ton-labs-node-tools/target/release/console" "${TOOLS_DIR}/"
-mv "${TOOLS_DIR}/keygen" "${TMP_DIR}/" &> /dev/null || true
-cp "${NODE_BUILD_DIR}/ton-node/ton-labs-node-tools/target/release/keygen" "${TOOLS_DIR}/"
+if [ -f "${NODE_BUILD_DIR}/ton-labs-node-tools/target/release/console" ]; then
+    mv "${TOOLS_DIR}/console" "${TMP_DIR}/" &> /dev/null || true
+    cp "${NODE_BUILD_DIR}/ton-labs-node-tools/target/release/console" "${TOOLS_DIR}/"
+fi
+if [ -f "${NODE_BUILD_DIR}/ton-labs-node-tools/target/release/keygen" ]; then
+    mv "${TOOLS_DIR}/keygen" "${TMP_DIR}/" &> /dev/null || true
+    cp "${NODE_BUILD_DIR}/ton-labs-node-tools/target/release/keygen" "${TOOLS_DIR}/"
+fi
 
 echo "INFO: build utils (ton-labs-node-tools)... DONE"
 
@@ -76,11 +82,13 @@ cd "${NODE_BUILD_DIR}/tonos-cli" && git checkout "${TONOS_CLI_GITHUB_COMMIT_ID}"
 cargo update
 cargo build --release
 
-mv "${TOOLS_DIR}/tonos-cli" "${TMP_DIR}" &> /dev/null || true
-cp "${NODE_BUILD_DIR}/tonos-cli/target/release/tonos-cli" "${TOOLS_DIR}/"
+if [ -f "${NODE_BUILD_DIR}/tonos-cli/target/release/tonos-cli" ]; then
+    mv "${TOOLS_DIR}/tonos-cli" "${TMP_DIR}" &> /dev/null || true
+    cp "${NODE_BUILD_DIR}/tonos-cli/target/release/tonos-cli" "${TOOLS_DIR}/"
+fi
 
 echo "INFO: build utils (tonos-cli)... DONE"
 
 echo "INFO: pull TON Labs contracts..."
-git clone https://github.com/tonlabs/ton-labs-contracts.git "${NODE_BUILD_DIR}/ton-labs-contracts"
+git clone https://github.com/tonlabs/ton-labs-contracts.git "${SRC_TOP_DIR}/ton-labs-contracts"
 echo "INFO: pull TON Labs contracts... DONE"
